requirements.txt
fastapi==0.135.2
GitPython==3.1.46
huggingface_hub==0.23.4
langchain_groq==1.1.2
langchain_text_splitters==1.1.1
pydantic==2.12.5
pytest==9.0.2
python-dotenv==1.2.2
qdrant_client==1.17.1
requests==2.33.1
sentence-transformers==5.2.3
streamlit==1.55.0
torch==2.2.2
torchvision==0.18.2
rank_bm25==0.2.2
transformers==4.41.2 
accelerate==0.30.1 